+++ packages/pds/src/lexicon/types/com/atproto/server/createAccount.ts
@@ -17,6 +17,7 @@ export interface InputSchema {
   inviteCode?: string
   password: string
   recoveryKey?: string
+  plcOp?: Uint8Array
   [k: string]: unknown
 }
 
@@ -25,6 +26,7 @@ export interface OutputSchema {
   refreshJwt: string
   handle: string
   did: string
+  didDoc?: {}
   [k: string]: unknown
 }