+++ packages/pds/tests/proxied/read-after-write.test.ts
@@ -21,7 +21,7 @@ describe('proxy read after write', () => {
     })
     agent = network.pds.getClient()
     sc = network.getSeedClient()
-    await basicSeed(sc)
+    await basicSeed(sc, { addModLabels: true })
     await network.processAll()
     alice = sc.dids.alice
     carol = sc.dids.carol